Added functionality to read text from a CSV that can be changed by the community.
Created MESSAGE_CUSTOM.C to handle the parsing of the file.
Added a function in MESSAGE.C to get the current language set in the game.
Added a cust_txt.csv file as part of the CommunityAssets.
Prepared Makefile to include the object of MESSAGE_CUSTOM.
Prepared C_EXTERN.H to include the external functions of MESSAGE_CUSTOM.