One MAKEFILE to rule them all
This changes so that SOURCES/MAKEFILE has built in dependencies to all
the other libraries it needs to create LBA0.EXE. It will automatically
run wmake in the directories where a library is missing.

It also removes the need for LINK.LD altogether. The (Open) Watcom C/C++
Tools User Guide states:

> Under DOS, an asterisk prefix (*) will cause Make to examine the length of the command argument. If it is
> too long (> 126 characters), it will take the command argument and stuff it into a temporary environment
> variable and then execute the command with "@env_var" as its argument.
...
> The command must, of course, support the "@env_var" syntax. Typically, DOS commands do not support
> this syntax but many of the Open Watcom tools do.

This is used to pass the full list of objects and libraries directly on
the command line to wlink, so a temporary file is no longer needed with
the linker directives.